requirements.txt
asgiref==3.8.1
boto3==1.37.18
botocore==1.37.18
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
jmespath==1.0.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
